x86/sgx: Pull enclave page max protections from SECINFO
The jury is still out on whether requiring userspace to define maximal
protections is the most desirable approach for enabling SGX2 interaction
with LSMs.  Given that the kernel doesn't yet support SGX2, pull the
maximal protections from the page's SECINFO flags so as not to
prematurely commit to uglifying the uapi, but reserve bytes in the
struct so that @prot can be added in the future with minimal fuss.

Note, this effectively blocks any potential use of ENCLU[EMODPE] until
the kernel supports SGX2.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>