x86/sgx: Add helpers to expose ECREATE and EINIT to KVM
Provide wrappers around __ecreate() and sgx_einit() to hide the ugliness
of overloading the ENCLS return value to encode multiple error formats
in a single int.  KVM will trap-and-execute ECREATE and EINIT as part
of virtualization SGX, and on an exception, KVM needs the trapnr so
that it can inject the correct fault to the guest.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>