x86/sgx: Define new ENCLS leafs related to VMM EPC oversubscription
Define new ENCLS leafs ERDINFO, ETRACKC, ELDBC and ELDUC.  ETRACKC and
ELD*C are variants of existing ENCLS leafs that return an error code
instead of generating a #GP if an EPC conflict is encountered.  These
variants enable a VMM to swap its VMs' pages in/out of the EPC without
generating faults in the VMM.  ERDINFO provides introspection
capabilities on EPC pages and is primarily intended for VMM use when
swapping out an EPC page, e.g. to determine if the page being swapped
should be reloaded in the BLOCKED or UNBLOCKED state.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>