x86/vdso: x86/sgx: Rework __vdso_sgx_enter_enclave() API
Rework __vdso_sgx_enter_enclave() to use a struct to hold the input and
output params.  In the new struct, add an opaque "user_data" that can be
used to pass context across the vDSO, an explicit "exit_reason" to avoid
overloading the return value, and a "flags" field to provide a path for
future extensions.

Moving the params into a struct will also make it less painful to use
dedicated exit reasons values in a future patch.

Cc: Nathaniel McCallum <npmccallum@redhat.com>
Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>