x86: Drop RDPID macro to discourage use of RDPID in the kernel
Drop the GAS-compatible RDPID macro as RDPID shouldn't be used in the
kernel, excepting the vDSO, which handcodes RDPID in inline asm.  RDPID
is unsafe in the kernel because KVM loads guest's TSC_AUX on VM-entry
and may not restore the host's value until the CPU returns to userspace.

See commit 6a3ea3e68b8a ("x86/entry/64: Do not use RDPID in paranoid
entry to accomodate KVM") for details.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>