KVM: x86: Remove obsolete disabling of page faults in kvm_arch_vcpu_put()
Remove the disabling of page faults across kvm_steal_time_set_preempted()
as KVM now accesses the steal time struct (shared with the guest) via a
cached mapping.  The cache lookup is flagged as atomic, thus it would
be a bug if KVM tried to resolve a new pfn, i.e. we want the splat that
would be reached via might_fault().

Signed-off-by: Sean Christopherson <seanjc@google.com>